asn1: allow missing optional raw values
For a field of type RawValue that is annotated with "optional,tag:N",
need to cope with that field not being present in the data.

The current code has a special case for RawValue that didn't cope
with this, so extend it; this also involved moving up the fragment
that extracts the expected class / tag information from the parameters.