examples/ct: fix verified chain check
The cert checker tests that the particular chain submitted on the
add-[pre-]chain request is a valid path to an accepted root.

The validated chain(s) returned from the X.509 library might be
different than that submitted, but the existing code had a bug
that meant that it effectively only checked that the chains
were of the same length, not that they were the same chain.