Add canonical methods to retrieve trees and acquire hashers/signers (#518)
* Add canonical methods to retrieve trees and acquire hashers/signers from them

GetTree retrieves a tree from either storage or context, with added validation.
Context retrieval is meant to be used for request-level trees. It may be used by
storage, servers and interceptors to retrieve trees without skipping the proper
checks.

Hasher and Signer methods are the canonical sources of how to acquire the
corresponding entities from trees.

* Review comments: rename NewSigner to NewSHA256Signer

* Review comments: docs and readability improvements