Replace ctfe.LogConfig struct with a similar protobuf
This is to enable using Protobuf "Any" messages in the config. In turn,
that makes it possible to replace the PrivKeyPEMFile/PrivKeyPassword/
PubKeyPEMFile fields in the config with one of the protos from
trillian/crypto/keyspb. This will lead to CTFEs being able to load their
keys from anywhere that Trillian can load keys from.