scanner,preload: remove channel buffers
The fetches chan buffer can be removed without any impact as the
generator is fast, and has to block on the consumer anyway.

The big arbitrary jobs buffer feels misplaced hidden in a generic-use
Scanner object, it should be up to the application to decide if it wants to
queue or not. It also unavoidably completely fills when the consumer is
slower than the fetcher (like in preload), adding 6GB of memory usage.

The point of buffering in preload is to not ever have a submitter block
on a fetch. Assuming a fetch takes at most X times a submit, and
considering fetches are batched while submits are not, a buffer of
length X (for each submitter) will suffice to guarantee that. I picked
10 as a conservative X, but I suspect it might be < 1.

Notice how preload can make a more informed decision on queueing than
scanner, confirming the choice about the jobs buffer.