trillian/ctfe: ignore unhandled PolicyConstraints
We have seen an occasional intermediate CA that has the
PolicyConstraints extension in its cert.  This extension
must be marked critical, but is not handled by the x509
parsing library, and so triggers a chain verification
failure.  Explicitly clear out the OID for this extension
from UnhandledCriticalExtensions to prevent this verification
failure.