x509: make nil CFDataRef types version-specific
Go 1.10 needs C.CFDataRef to have nil value of '0'
Go 1.9 nees C.CFDataRef to have nil value of 'nil'

As we have a fork of crypto/x509, our source code for x509 can be
a different version than the current compiler.

We need our code to work with both 1.9 and 1.10, so encapsulate
setting and testing for nil CFDataRef into a couple of
version-specific functions.

Fixes #153 and #131 (hopefully).