trillian/ctfe: MaxPathLen++ for preissued precerts
The x509.Verify() entrypoint will check that any MaxPathLen constraints
for CA certificates in the chain are respected; however, these limits
will not allow for the presence of an additional intermediate pre-isser
CA certificate.  Add one across the board to allow for this (and
remember to take it off again afterwards).