Add CORS handler to ct_server
This CORS handler will allow cross-origin requests to log URLs and for
health checking. Other URLs, such as metrics, will not allow cross-origin
requests. This is safe because CT logs are public and unauthenticated so
cross-site scripting attacks are not a concern. It will enable Javascript
clients running in browsers to access CT logs.