types: Add JSON tags to `SignedCertificateTimestamp`.
Prior to this commit the `ct.SignedCertificateTimestamp` object is
missing `json` tags that would allow it to be marshaled to JSON.
This is useful for (as one example) providing
a `ct.SignedCertificateTimestamp` instance as the result of a call to
the "add-chain" method of a CT server that (per RFC 6962, 4.1) should
return a JSON encoded SCT.

This commit adds JSON tags to the `ct.SignedCertificateTimestamp` type
and includes a small unit test to ensure the marshal works as expected.
This saves users of the library from having to make their own throw-away
SCT type in order to add JSON tags for marshaling.