x509: add IsFatal(err) helper and use in tests
This fork differs from the upstream crypto/x509 in its error
behaviour:
 - Parsing may return a certificate *and* a non-nil error of type
   NonFatalErrors
 - There are additional checks that are not present in the upstream
   code.

To make it easier to cope with this, add an IsFatal(error) function
to identify fatal errors, and move various test cases that use
technically-invalid certificates to use this check.

The revoked.go code also uses a more detailed Error/Errors type; cope
with this too.