Fix golang 1.11 compatibility on Windows
Go 1.11beta1 introduces an incompatible change to the type of
CertChainPolicyPara.ExtraPolicyPara field [1], which results
in the following compile error on Windows:

> x509\root_windows.go:112:3: cannot use
> uintptr(unsafe.Pointer(sslPara)) (type uintptr) as type
> syscall.Pointer in field value

The fix, maintaining backward compatibility with Go 1.10 and 1.9,
is to use type alias. Unfortunately it won't work with go < 1.9
as those earlier versions lack the type alias feature.

This should fix issue #284.

[1] https://github.com/golang/go/commit/4869ec00e87ef

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>