x509: make Windows policy parameter type version-specific
Go 1.11 needs syscall.CertChainPolicyPara.ExtraPolicyPara to
have type syscall.Pointer, but in previous versions of Go this
had type uintptr.

As we have a fork of crypto/x509, our source code for x509 can be
a different version than the current compiler.

To allow our code to work with both 1.11 and earlier versions,
encapsulate the cast into a version-specific function.

Fixes #284