Introduce ValidatedLogConfig (#308)
This change addresses the issue of having to parse LogConfig at least 2 times:
one for validating the config, and one for creating the log instance. The key
idea is to return the parsed objects from the validation function, and re-use
them in subsequent phases.

Additionally, this change moves the ExtKeyUsage checks to ValidateLogConfig, to
be returned in the ValidatedLogConfig as well.