Factor out TLS parsing of LeafEntry (#318)
This change allows checking whether the format of get-entries is compliant with TLS encoding specified in RFC6962, independently from parsing x509.

This helps to solve the following problem in CT mirroring. Different logs use different x509 parsers, so when Migrillian populates a mirror it might encounter a certificate/chain that it can't parse. But it seems that mirrors should be able to let such data through. To make admission control laxer Migrillian can check only the TLS encoding (but not too lax, so it should in addition make sure the Merkle Tree leaf hash is consistent with the tree).