Use specific tag for Docker image in Kubernetes config
The "latest" flag is just whatever the most recently pushed image is, which
might not actually be the most up-to-date. It also results in automatic updates
of the image used in the Kubernetes cluster, which might be unexpected. It's
better to use a specific tag (e.g. a commit hash or release version). This makes
it obvious which version of the code is being used in Kubernetes.