PostAndParseWithRetry: Return RspError for status mismatch.
Prior to 5b1f90945db965133564fdfbe129c018e6150742 the client callers of
`jsonclient.PostAndParseWithRetry` would convert a non-nil err response
into a `RspError` when the `httpRsp` was also non-nil. After the
refactor the result from `PostAndParseWithRetry` is returned directly
and it is assumed that function will return a `RspError` in all
cases.

There exists one case where a bare `fmt.Errorf` was being
returned instead of an `httpRsp`, breaking this assumption and masking
the HTTP response body from interested callers to
`PostAndParseWithRetry`.

This commit updates the `PostAndParseWithRetry` unit test to ensure when
a non-nil error is returned it is always a `RspError` instance or
`context.DeadlineExceeded`. Without the fix to
`PostAndParseWithRetry` the updated test fails as expected:

```
go test --test.run TestPostAndParseWithRetry ./jsonclient/...
2018/11/26 14:19:13 Request failed, backing-off on http://127.0.0.1:39638 for 0s: parse http://127.0.0.1:39638/short%: invalid URL escape "%"
2018/11/26 14:19:13 Request timed out, retrying immediately
2018/11/26 14:19:13 Request failed, backing-off for 0s: got HTTP status 503 Service Unavailable
2018/11/26 14:19:13 Request failed, backing-off for 0s: got HTTP status 503 Service Unavailable
--- FAIL: TestPostAndParseWithRetry (0.00s)
    client_test.go:441: PostAndParseWithRetry()=*errors.errorString; want jsonClient.RspError
    FAIL
    FAIL  github.com/google/certificate-transparency-go/jsonclient  0.005s
```

After the fix to `PostAndParseWithRetry` the updated test passes.