[Migrillian] Add option for skipping consistency check (#444)
Some CT logs in the ecosystem don't expose get-entry-and-proof endpoint
which makes it impossible for Migrillian to check partial consistency
between Trillian pre-ordered log and the source log. This change adds a
config option allowing to opt out from these checks for such logs.

Note that get-entry-and-proof is not an optional endpoint though, and the
logs that don't have it are actually non-compliant.