[Migrillian] Fix deadlock when terminating submitters
When shutdown happens it's possible that submitters stop consuming items
from the fetchers->submitters channel, and some fetchers continue trying
to push to this channel. If the channel happens to be full then the
whole pipeline deadlocks. This PR fixes the problem by failing the
channel write if the context is being canceled.