If realip is determined to be empty string, return clearer sentinel
Empty string is a very unclear way of showing the address wasn't found. This should be a safe change, the only potential risk is that during rollout of this version compared to the previous version, any requests that previously were mapping to emptry string and were being limited will now get mapped to an empty bucket and get access to a one-time freebie.