Factor out conversion from AddChainResponse to SCT.
This is mostly extracted from what is done in
LogClient.addChainWithRetry, but with a stronger check on the length of
the id field in the AddChainResponse.  Unfortunatly this additional check makes
it non-trivial to refactof LogClient.addChainWithRetry, as the SCT used
in the tests for that function use an SCT with an empty ID field.  Thus,
that refactoring will come in a follow up PR.