Fix Sharding logic with using notbefore of a certificate.
According to https://github.com/chromium/ct-policy/issues/6, a NotAfter
validity field of a X509 Certificate should be used.

The policy is as follows.
For a certificate to be accepted by the Log that has a time range
specified.
The certificate’s ‘Not After’ validity field value must:
Be at or after ‘Start’ and Be before ‘End’