[witness] adding log ID (#875)
When parsing an STH, the witness now checks if its LogID field matches its own idea of what the log ID should be and either returns an error (if they don't match), leaves it alone (if they do match), or fills it in if it was empty. This makes it easier for a client who is given a (cosigned) STH to know which log public key to use to verify.