Move PEMCertPool from CTFE package to x509util
Currently many things indirectly depend on the entire CTFE package
(including HTTP handlers and some flags) because they use the cert pool.
For example, the ctclient CLI tool imports loglist, which imports ctfe.
This is a large overkill, and brings irrelevant flags to the CLI tool.

This change moves the cert pool type to x509util, because it's a wrapper
around the x509 pool.