Replace juju/ratelimit with x/time/rate (#800)
The consequences of the LGPL static linking exception in juju/ratelimit
aren't terribly well understood, and it's the only component in the
certificate-transparency chain that's not under something well-studied like
MIT or Apache. There's now ratelimit support in x/time, so porting over to
use that instead simplifies licensing analysis.