ctfe: Add readonly mode (#927)
The new is_readonly flag in CTFE config allows setting up logs that accept
only read requests. This is particularly useful in the scenario when a CT log is
gradually phased out from the ecosystem.

Handling readonly transition at CTFE level is beneficial over (in some ways),
but is mostly complementary to the current approach of setting the Trillian log
to DRAINING / FROZEN state.

The Trillian state transition is:

- more responsive (it happens immediately after the state update RPC), and
- more defensive w.r.t. correctness (it prevents new writes to the Merkle tree
on Trillian level).

The state transition at the CTFE level is:

- as [not] responsive as the config rollout procedure, and
- less strict in a sense that updates might still slip through to the Merkle
tree (e.g. if some CTFE job is stale or rolled back), but is
- more resource-efficient, as it rejects write requests early.