Enable all linters for ct-go (#1064)
Fixed lint errors in best local way possible. Some of these cases seem like they should be propagated further, but that is outside the scope of this work and requires more context than I can load in now. The result of these changes is that errors should at least no longer silently be dropped.

There are still some lint rules that are not strictly enforced by this change, specifically https://golangci-lint.run/usage/false-positives/#exc0001. These are enforced by the trillian and trillian-examples repos. We should aim to get consistency here, but change represents a positive step in the right direction whether we choose to live with this exception or remove it and fix the remaining errors.